Initial Commit; Basic Mecanum Drive Code
Just creates a WPILIB MecanumDrive object, and then loops in Teleop to pass it the Xbox Controller input. Also has a SpeedModifier Variable which defaults to 50% and is changeable either in code, or with the Left and Right bumpers while in Teleop. (State will be printed on bumper press) Finally, creates a Autonomous selector for Java SmartDashboard. No Auto code yet.